avance 2

+++ Proyecto Final - QR Parking/Clases/Tablas/TablaPersonalRegistro.cs
@@ -70,9 +70,13 @@ namespace Proyecto_Final___QR_Parking.Clases.Tablas
 						writer.Write(registro.Contra);
 					}
 
-					File.Delete(nombre_fichero);
-					File.Move(temp, nombre_fichero);
 				}
+
+				if (File.Exists(nombre_fichero))
+					File.Delete(nombre_fichero);
+
+				File.Move(temp, nombre_fichero);
+				return true;
 			}
 
 			catch (Exception ex)
@@ -86,22 +90,13 @@ namespace Proyecto_Final___QR_Parking.Clases.Tablas
 
 				return false;
 			}
-			
-			return true;
 		}
 
 		public void CargarPersonal()
 		{
 			if (!File.Exists(nombre_fichero))
 			{
-				MessageBox.Show(
-					"¡Aún no ha guardado ningún registro de Personal de Registros!",
-					"Advertencia",
-					MessageBoxButtons.OK,
-					MessageBoxIcon.Warning
-				);
-				
-				return;
+				File.Create(nombre_fichero).Close();
 			}
 
 			using (var stream = File.Open(nombre_fichero, FileMode.Open, FileAccess.Read))
