+++ Proyecto Final - QR Parking/Clases/Tablas/TablaSeguridad.cs
@@ -69,10 +69,13 @@ namespace Proyecto_Final___QR_Parking.Clases.Tablas
 						writer.Write(registro.Correo);
 						writer.Write(registro.Contra);
 					}
-
-					File.Delete(nombre_fichero);
-					File.Move(temp, nombre_fichero);
 				}
+
+				if (File.Exists(nombre_fichero))
+					File.Delete(nombre_fichero);
+
+				File.Move(temp, nombre_fichero);
+				return true;
 			}
 
 			catch (Exception ex)
@@ -87,21 +90,13 @@ namespace Proyecto_Final___QR_Parking.Clases.Tablas
 				return false;
 			}
 
-			return true;
 		}
 
 		public void CargarPersonal()
 		{
 			if (!File.Exists(nombre_fichero))
 			{
-				MessageBox.Show(
-					"¡Aún no ha guardado ningún registro de Seguridad!",
-					"Advertencia",
-					MessageBoxButtons.OK,
-					MessageBoxIcon.Warning
-				);
-
-				return;
+				File.Create(nombre_fichero).Close();
 			}
 
 			using (var stream = File.Open(nombre_fichero, FileMode.Open, FileAccess.Read))
